Organize data for compatibility with Movebank
Data now able to be imported from Vectronic files, merged, cleaned up, and formatted for upload to Movebank, including creating a metadata file that identifies collar id, animal id, and deployment details, which is necessary for Movebank (data upload and metadata upload). Then, working to clean up the data and export in Movebank format, which is the starting point of many of the AniMove strategies discussed, i.e. helps keep things standardized.